telegram-bot/requirements.txt
python-telegram-bot==20.7
anthropic==0.40.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic-settings==2.1.0
loguru==0.7.2